backend/requirements.txt
fastapi
uvicorn
fastapi-mail
python-dotenv